pyproject.toml
[project]
name = "gpt-graphic-design-evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shunsuke KITADA", email = "shunsuke.kitada.0831@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "langchain[openai]>=0.3.0",
    "pillow>=12.0.0",
]

[project.optional-dependencies]
api = []

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.0.0",
    "pytest>=6.0.0",
    "ruff>=0.1.5",
]
